fix: move file parsing from init to analyze
