+++ packages/cli/src/types.ts
@@ -14,16 +14,11 @@ export type MigrateCommandOptions = {
   format: Formats[];
   outputPath: string;
   verbose: boolean | undefined;
+  userConfig: string | undefined;
   ci: boolean | undefined;
   dryRun: boolean;
   regen: boolean | undefined;
 };
-
-// export type MigrateInitCommandOptions = {
-//   basePath: string;
-//   verbose: boolean | undefined;
-//   userConfig: string | undefined;
-// };
 
 export type MigrateCommandContext = {
   skip: boolean;
